[debops.sshd] Fix sshd macs ,ciphers and kex_algorithms generation
The for loop contains a test to avoid more than one iteration.
This has no effect as the "iterations" are evaluated at once.

This issue is documented in http://jinja.pocoo.org/docs/2.10/templates/#for
"Please note that assignments in loops will be cleared at the end of the
iteration and cannot outlive the loop scope. Older versions of Jinja2
had a bug where in some circumstances it appeared that assignments would
work. This is not supported. See Assignments for more information about
how to deal with this.

More details and alternatives constructs at:
http://jinja.pocoo.org/docs/2.10/templates/#assignments