[debops.sshd] Fix off-by-one error in moduli tasks
The default value for 'sshd__moduli_minimum' is 2048. Looking in
'/etc/ssh/moduli' file on a DebOps prepared system, the smallest size
entries are 3071 and the vanilla file in the Buster package contains
smallest entries with a size of 2047.

The modulus string is 512 byte long, which is a hexadecimally encoded
number, decoded to a number of 256 byte (2048 bit). Therefore DebOps
should filter the entries with a reported size <= 'sshd__moduli_minimum' - 1.